refactor(PhotoParams): 별도 타입으로 분리
