+++ src/app/api/background/route.ts
@@ -1,25 +1,6 @@
+import type { PhotoParams } from '@/entities/background';
 import { type NextRequest, NextResponse } from 'next/server';
-import {
-  type ColorId,
-  type ContentFilter,
-  createApi,
-  type Language,
-  type Plus,
-  type SearchOrderBy
-} from 'unsplash-js';
-import { PaginationParams, OrientationParam } from 'unsplash-js/dist/types/request';
-
-export type PhotoParams = {
-  query: string;
-} & Pick<PaginationParams, 'page' | 'perPage'> &
-  OrientationParam & {
-    orderBy?: SearchOrderBy;
-    color?: ColorId;
-    plus?: Plus;
-    lang?: Language;
-    contentFilter?: ContentFilter;
-    collectionIds?: string[];
-  };
+import { createApi } from 'unsplash-js';
 
 export async function GET(request: NextRequest) {
   const { searchParams } = request.nextUrl;
