+++ src/entities/background/model/types.ts
@@ -1,3 +1,6 @@
+import type { SearchOrderBy, ColorId, Plus, Language, ContentFilter } from 'unsplash-js';
+import type { PaginationParams, OrientationParam } from 'unsplash-js/dist/types/request';
+
 export interface BackGroundPhotoInstance {
   results: BackGroundPhoto[];
   total: number;
@@ -14,3 +17,15 @@ export interface BackGroundPhoto {
   };
   width: number;
 }
+
+export type PhotoParams = {
+  query: string;
+} & Pick<PaginationParams, 'page' | 'perPage'> &
+  OrientationParam & {
+    orderBy?: SearchOrderBy;
+    color?: ColorId;
+    plus?: Plus;
+    lang?: Language;
+    contentFilter?: ContentFilter;
+    collectionIds?: string[];
+  };
